Refactor the members factory/command for more generic use
Fixes #111
Refs #116
Fixes #112
